\name{airquality}
\title{New York Air Quality Measurements}
\usage{data(airquality)}
\alias{airquality}
\format{A data frame with 154 observations on 6 variables.
  \tabular{rlll}{
    [,1]  \tab Ozone    \tab numeric  \tab Ozone (ppb) \cr
    [,2]  \tab Solar.R  \tab numeric  \tab Solar R (lang) \cr
    [,3]  \tab Wind     \tab numeric  \tab Wind (mph) \cr
    [,4]  \tab Temp     \tab numeric  \tab Temperature (degrees F) \cr
    [,5]  \tab Month    \tab numeric  \tab Month 1-12 \cr
    [,6]  \tab Day      \tab numeric  \tab Day of month 1-31 \cr
    }
}
\source{The data were obtained from the New York State Department of
Conservation (ozone data) and the National Weather Service
(meteorological data).
}
\description{Daily readings of the following air quality values for
May 1, 1973 (a Tuesday) to September 30, 1973: mean ozone in parts per
billion from 1300 to 1500 hours at Roosevelt Island, solar radiation
(Solar.R) in Langleys in the frequency band 4000-7700 Angstroms from
0800 to 1200 hours at Central Park, Average wind speed (Wind) in miles
per hour at 0700 and 1000 hours at LaGuardia Airport, maximum daily
temperature (Temp) in degrees Fahrenheit at La Guardia Airport.

}
\references{Chambers, J. M., W. S. Cleveland, B. Kleiner and P. A. Tukey (1983).
Graphical Methods for Data Analysis.
Belmont, CA: Wadsworth.

}
\keyword{datasets}
