\name{anscombe}
\title{Anscombe's Quartet of ``Identical'' Simple Linear Regressions}
\usage{data(anscombe)}
\alias{anscombe}
\format{A data frame with 11 observations on 8 variables.
  \tabular{rl}{
   x1 == x2 == x3 \tab the integers 4:14, specially arranged \cr
   x4 \tab values 8 and 19 \cr
   y1, y2, y3, y4 \tab numbers in (3, 12.5)\cr
                  \tab mean(y<j>) = 7.5 and st.dev = 2.03
    }


}
\source{Edward R. Tufte (1989).  The Visual Display of Quantitative Information;
	Graphics Press, p.13-14.
}

\references{
Francis J. Anscombe (1973). Graphs in Statistical Analysis;
	American Statistician, 27, 17-21.
}
\description{Four x-y datasets which have the same traditional statistical properties
(mean, variance, correlation, regression line, etc.), yet are quite
different.
}
\keyword{datasets}
