% File src/library/base/man/base-package.Rd
% Part of the R package, http://www.R-project.org
% Copyright 1995-2007 R Core Team
% Distributed under GPL 2 or later

\name{base-package}
\alias{base-package}
\alias{base}
\docType{package}
\title{
The R Base Package
}
\description{
Base R functions
}
\details{
This package contains the basic functions which let \R{}
function as a language:  arithmetic, input/output, basic
programming support, etc.  Its contents are available through
inheritance from any environment.

For a complete list of functions, use \code{library(help = "base")}.
}
\keyword{ package }
