\name{base.package}
\alias{base.package}
\alias{base}
\docType{package}
\title{
The R Base Package
}
\description{
Base R functions
}
\details{
This package contains the basic functions which let \R{}
function as a language:  arithmetic, input/output, basic
programming support, etc.  Its contents are available through
inheritance from any environment.
}
\keyword{ package }
