% File src/library/base/man/capture.Rd
% Part of the R package, https://www.R-project.org
% Copyright 2016 R Core Team
% Distributed under GPL 2 or later

\name{capture}
\title{Capture arguments}
\usage{
captureArg(x)
captureDots()
}
\alias{captureArg}
\alias{captureDots}
\description{
}
\arguments{
  \item{x}{a symbol representing a function argument.}
}
\details{
}
\value{
}
\note{
}
\seealso{
  \code{\link{substitute}} for argument substitution.
}
\examples{
NULL
}
\keyword{programming}
