\name{cars}
\title{Stopping Distances of Cars}
\usage{data(cars)}
\alias{cars}
\format{A data frame with 50 observations on 2 variable.
  \tabular{rlll}{
    [,1]  \tab speed  \tab numeric  \tab Speed (mph)\cr
    [,2]  \tab dist   \tab numeric  \tab Stopping distance (ft)\cr
  }
}
\source{Ezekiel, M. (1930).  Methods of Correlation Analysis.  Wiley.
}
\description{The data give the speed of cars and the distances taken to stop.
Note that the data were recorded in the 1920s.

}
\references{McNeil, D. R. (1977).  Interactive Data Analysis.  Wiley.

}
\keyword{datasets}
