\name{character}
\title{Character String Vectors}
\usage{
character(n=0)

as.character(x)
is.character(x)
}
\alias{character}
\alias{as.character}
\alias{is.character}
\value{
This function creates a character vector of the specified length.
The elements of the vector are all equal to \code{""}.

\code{as.character} attempts to coerce its argument to be of character
type.

\code{is.character} returns \code{TRUE} or \code{FALSE} depending on
whether its argument is of character type or not.
}
\keyword{character}
\keyword{classes}
