\name{chickwts}
\title{Chicken Weights by Feed Type}
\usage{data(chickwts)}
\alias{chickwts}
\format{A data frame with 71 observations on 2 variables.
  \tabular{rlll}{
    [1,]  \tab weight  \tab numeric  \tab Chick weight \cr
    [2,]  \tab feed    \tab factor   \tab Feed type
  }
}
\source{Anonymous (1948).  Biometrika, p.214.
} 
\description{An experiment was conducted to measure and compare the
  effectiveness of various feed supplements on the growth rate of
  chickens.  Newly hatched chicks were randomly allocated into six groups,
  and each group was given a different feed supplement.  Their weights in
  grams after six weeks are given along with feed types.
}
\references{McNeil, D. R. (1977).  Interactive Data Analysis.  New York: Wiley.
}
\keyword{datasets}
