\name{colors}
\title{Color Names}
\usage{
colors()
}
\alias{colors}
\value{
This function returns the built-in color names which R knows about.
These names can be used with a \code{col=} specification in graphics
functions.

An even wider variety of colors can be created with \code{rainbow},
\code{rgb} and \code{hsv}.
}
\seealso{
\code{\link{rainbow}} for a nice example;
\code{\link{rgb}}, \code{\link{hsv}}, \code{\link{gray}}, 
and \code{\link{heat.colors}}, \code{\link{topo.colors}} for images.
}
