\name{contributors}
\title{R Project Contributors}
\alias{contributors}
\section{The \R Project}{
  \R is a project which is attempting to provide a modern
  piece of statistical software to the GNU suite of software.
}
\section{Contributors}{
  The current \R is the result of a collaborative effort of several
  authors from all over the world.
  \R was initially written by Robert Gentleman and Ross Ihaka ---
  also known as ``\emph{R & R}'' ---
  of the Statistics Department of the University of Auckland.

  Since mid-1997 there has been a core group with access to the \R
  source, currently consisting of

  \itemize{%%-- should we add E-mail  and / or affiliation ?
      \item Douglas Bates
      \item Peter Dalgaard
      \item Robert Gentleman
      \item Kurt Hornik
      \item Ross Ihaka
      \item Friedrich Leisch
      \item Thomas Lumley
      \item Martin Maechler
      \item Paul Murrell
      \item Heiner Schwarte
      \item Luke Tierney
    }

  In addition, a large group of individuals has contributed to \R by
  sending code and bug reports, notably

  Valerio Aimalea,
  Ben Bolker,
  Paul Gilbert,
  Philippe Lambert
  Jim & Patrick Lindsey,
  Guido Masarotto,
  Steve Oncley and Gordon Maclean,
  Hubert Palme,
  Martyn Plummer,
  Brian Ripley,
  Bill Simpson,
  Bill Venables,
  Andreas Weingessel
}
\keyword{misc}
