% File src/library/base/man/copyright.Rd
% Part of the R package, http://www.R-project.org
% Copyright 1995-2007 R Core Team
% Distributed under GPL 2 or later

\name{copyright}
\title{Copyrights of Files Used to Build R}
\alias{copyright}
\alias{copyrights}
\description{
  \R is released under the \sQuote{GNU Public License}: see
  \code{\link{license}} for details.  The license describes your right
  to use \R.  Copyright is concerned with ownership of intellectual
  rights, and some of the software used has conditions that the
  copyright must be explicitly stated: see the \sQuote{Details} section.  We
  are grateful to these people and other contributors (see
  \code{\link{contributors}}) for the ability to use their work.
}
\details{
  The file \file{\var{\link{R_HOME}}/COPYRIGHTS} lists the copyrights in full
  detail.
}
\keyword{misc}
