\name{Defunct}
\title{Defunct Functions}
\alias{Defunct}
\alias{.Defunct}
\usage{
.Defunct()

%--- oldfun(.)
category(x, levels, labels, ordered, exclude)
}
%------ PLEASE: put \alias{.} here for  EACH !
\alias{category}
%
\description{
  The functions listed here are no longer part of \R as they are not
  needed (any more):

  \code{category} has been an old-S function before there were factors;
  should be replaced by \code{\link{factor}} throughout!

  \code{.Defunct} is the function to which defunct functions are set.
}
\seealso{\code{\link{Deprecated}}}
\keyword{documentation}
\keyword{utilities}

