\name{eurodist}
\alias{eurodist}
\title{Distances Between Cities in Europe}
\usage{data(eurodist)}
\description{
  The data give the road distances (in km) between 21 cities in Europe.
  The data are taken from a table in ``The Cambridge Encyclopaedia''.
}
\format{
  A \code{dist} object based on 21 objects.  (You must have the mva
  package loaded to have the methods for this kind of object available).
}
\source{
  Crystal, D. Ed. (1990)  \emph{The Cambridge Encyclopaedia.}
  Cambridge: Cambridge University Press,
}
\keyword{datasets}
