\name{fivenum}
\title{Tukey Five-Number Summaries}
\usage{
fivenum(x)
}
\alias{fivenum}
\value{
A five number summary (minimum, lower-hinge, median,
upper-hinge, maximum) for the numbers in \code{x}
}
\seealso{
\code{\link{IQR}}, \code{\link{median}}, \code{\link{quantile}},
\code{\link{range}}.
}
\keyword{univar}
\keyword{robust}
\keyword{distribution}

