\name{Formaldehyde}
\title{Determination of Formaldehyde}
\usage{data(Formaldehyde)}
\alias{Formaldehyde}
\format{A data frame with 6 observations on 2 variables.
    \tabular{rlll}{
    	[,1] \tab carb\tab numeric \tab Carbohydrate (ml) \cr
    	[,2] \tab opt \tab numeric \tab Optical Density \cr
  }
}
\source{Bennett, N. A. and N. L. Franklin (1954).  Statistical analysis
    in Chemistry and the Chemical Industry.  New York: Wiley
}
\description{These data are from a chemical experiment to prepare a
    standard curve for the determination of formaldehyde by the addition
    of chromatropic acid and concentrated sulpuric acid and the reading
    of the resulting purple color on a spectophotometer.
}
\references{
    McNeil, D. R. (1977).  Interactive Data Analysis.  New York: Wiley.
}
\keyword{datasets}
