\name{freeny}
\title{Freeny's Revenue Data}
\usage{data(freeny)}
\alias{freeny}
\format{
  \tabular{ll}{
    y: \tab quarterly revenue, 39 observations from (1962,2Q) to (1971,4Q).\cr
    x: \tab matrix of explanatory variables.  Columns are y lagged 1 \cr
       \tab quarter, price index, income level, and market potential.
     }
}
\source{A. E. Freeny (1977). A Portable Linear Regression Package with Test Programs.
Bell Laboratories memorandum.
}
\keyword{datasets}
