\name{graphics.off}
\title{Turn Off All Graphics Devices}
\usage{
graphics.off()
}
\alias{graphics.off}
\description{
\code{graphics.off()} shuts down all open graphics devices.
Use \code{\link{dev.off}()} to only close the current one.
}
\seealso{
\code{\link{dev.off}} and the other \code{dev.xxx} functions;
\code{\link{Devices}}.
}
\keyword{device}
\keyword{iplot}
