% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-theme.R
\name{guide_continuous}
\alias{guide_continuous}
\title{A nicely styled colorbar for continuous aesthetics (color, fill and alpha)}
\usage{
guide_continuous(aes = "color", type = "bar", width = 0.4, ...)
}
\arguments{
\item{aes}{(type) Type of the guide. Options include "bar" and "steps".}

\item{width}{(number) Width of the color bar.
Options include "color", "fill", "shape" and "size".}

\item{...}{Other arguments passed to ggplot methods.}
}
\description{
A nicely styled colorbar for continuous aesthetics (color, fill and alpha)
}
\examples{
\dontrun{
library(ggplot2)
ggplot(mpg, aes(cty, hwy)) + geom_point(aes(color = displ)) +
  theme_omics() + guide_continuous()
ggplot(mpg, aes(class)) + geom_bar(aes(fill = manufacturer)) +
  theme_omics() + guide_continuous(aes = "fill")
}

}
