% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-theme.R
\name{guide_discrete}
\alias{guide_discrete}
\title{Discrete guide}
\usage{
guide_discrete(aes = "color", reverse = FALSE, ...)
}
\arguments{
\item{aes}{(string) Aesthetic of the legend that should be modified.
Options include "color", "fill", "shape", "size" and "alpha".}

\item{reverse}{(boolean) If TRUE legend items are shown in reversed order.
Defaults to FALSE.}

\item{...}{Other arguments passed to ggplot methods.}
}
\description{
A nicely styled legend for categorical aesthetics (color, fill, shape,
size and alpha)
}
\examples{
\dontrun{
library(ggplot2)
ggplot(mpg, aes(cty, displ)) + geom_point(aes(color = manufacturer)) +
  theme_omics() + guide_discrete()
ggplot(mpg, aes(class)) + geom_bar(aes(fill = manufacturer)) +
  theme_omics() + guide_discrete(aes = "fill")
}

}
