% File src/library/base/man/identity.Rd
% Part of the R package, https://www.R-project.org
% Copyright 1995-2015 R Core Team
% Distributed under GPL 2 or later

\name{identity}
\alias{identity}
\title{Identity Function}
\description{
  A trivial identity function returning its argument.
}
\usage{
identity(x)
}
\seealso{
  \code{\link{diag}} creates diagonal matrices, including identity ones.
}
\arguments{
  \item{x}{an \R object.}
}
\keyword{programming}
