\name{image}
\title{Display a color image}
\usage{
image(x, y, z, zlim, col = heat.colors(12), ...)
}
\alias{image}
\arguments{
  \item{x,y}{locations of grid lines at which the values in \code{z} are
    measured.  These must be in ascending order.  By default, equally
    spaced values from 0 to 1 are used.}
  \item{z}{a matrix containing the values to be plotted (\code{NA}s are
    allowed).}
  \item{zlim}{the minimum and maximum \code{z} values for which colors
    will be plotted.  Each of the given colors will be used to color an
    equispaced interval of this range.} 
  \item{col}{A list of colors such as that generated by
    \code{\link{rainbow}}, \code{\link{heat.colors}},
    \code{\link{topo.colors}}, \code{\link{terrain.colors}} or similar
    functions.} 
  \item{\dots}{graphical parameters for \code{\link{plot}} may also be
    passed as arguments to this function.}
}
\description{
  Creates a grid of colored or gray-scale rectangles with colors
  corresponding to the values in \code{z}.  This can be used to display
  three-dimensional or spatial data. 
  The functions \code{\link{heat.colors}}, \code{\link{terrain.colors}}
  and \code{\link{topo.colors}} create heat-spectrum (red to white) and
  topographical color schemes suitable for displaying ordered data, with
  \code{n} giving the number of colors desired.
}
\note{
  Based on a function by Thomas Lumley
  \email{thomas@biostat.washington.edu}.
}
\seealso{
  \code{\link{contour}},
  \code{\link{heat.colors}},
  \code{\link{topo.colors}},
  \code{\link{terrain.colors}},
  \code{\link{rainbow}},
  \code{\link{hsv}},
  \code{\link{par}}.
}
\examples{
data(volcano)
x <- 10*(1:nrow(volcano))
y <- 10*(1:ncol(volcano))
image(x, y, volcano, col = terrain.colors(100), axes = FALSE)
contour(x, y, volcano, levels = seq(90, 200, by=5), add = TRUE, col = "peru")
axis(1, at = seq(100, 800, by = 100))
axis(2, at = seq(100, 600, by = 100))
box()
title(main = "Maunga Whau Volcano", font.main = 4)
}
