\name{insects}
\title{Effectiveness of Insect Sprays}
\usage{data(insects)}
\alias{insects}
\format{A data frame with 72 observations on 2 variables.
  \tabular{rlll}{
    [,1]  \tab count  \tab numeric  \tab Insect Count\cr
    [,2]  \tab spray  \tab factor   \tab The type of spray
  }
}
\source{Geoffrey Beall, (1942).  "The Transformation of Data from
Entomological Field Experiments", Biometrika,  pages 243-262.
}
\references{
McNeil, D. (1977).  Interactive Data Analysis.  New York: Wiley.
}
\description{The counts of insects in agricultural experimental units
treated with different insecticides.
}
\keyword{datasets}
