\name{is.function}
\alias{is.function}
\title{Is an Object of Type Function?}
\usage{is.function(x)}
\description{Checks whether its argument is a function.}
\arguments{
  \item{x}{an \R object.}
}
\details{
  \code{is.function} is generic: you can write methods to handle
  specific classes of objects, see \link{InternalMethods}.
}
\value{\code{TRUE} if \code{x} is a function, and \code{FALSE}
  otherwise.}
}
\keyword{programming}
