% File src/library/base/man/is.single.Rd
% Part of the R package, https://www.R-project.org
% Copyright 1995-2007 R Core Team
% Distributed under GPL 2 or later

\name{is.single}
\title{Is an Object of Single Precision Type?}
\usage{is.single(x)}
\alias{is.single}
\description{
  \code{is.single} reports an error.  There are no single precision
  values in R.
}
\arguments{
  \item{x}{object to be tested.}
}
\references{
  \bibshow{R:Becker+Chambers+Wilks:1988}
}
\keyword{classes}
