\name{judges}
\title{Lawyer's rating's of State Judges in the Superior Court.}
\usage{data(judges)}
\alias{judges}
\format{A data frame contaning 43 observations on 12 variables.
  \tabular{rlll}{
    [,1]  \tab CONT  \tab numeric  \tab Number of contacts \cr
    \tab \tab \tab of lawyer with judge. \cr
    [,2]  \tab INTG  \tab numeric  \tab Judicial integrity.\cr
    [,3]  \tab DMNR  \tab numeric  \tab Demeanor.\cr
    [,4]  \tab DILG  \tab numeric  \tab Diligence.\cr
    [,5]  \tab CFMG  \tab numeric  \tab Case flow managing.\cr
    [,6]  \tab DECI  \tab numeric  \tab Prompt decisions.\cr
    [,7]  \tab PREP  \tab numeric  \tab Preparation for trial.\cr
    [,8]  \tab FAMI  \tab numeric  \tab Familiarity with law.\cr
    [,9]  \tab ORAL  \tab numeric  \tab Sound oral rulings.\cr
    [,10]  \tab WRIT  \tab numeric  \tab Sound written rulings.\cr
    [,11]  \tab PHYS  \tab numeric  \tab Physical ability.\cr
    [,12]  \tab RTEN  \tab numeric  \tab Worthy of retention.\cr
  }
}
\source{New Haven Register, 14 January, 1977 (from John Hartigan).

}
\keyword{datasets}




