\name{license}
\title{Show the License Terms for R}
\usage{license()}
\usage{licence()}
\alias{license}
\alias{licence}
\description{
  The function just prints a few lines telling about the GNU license
  terms, and how to obtain more details.
}
\references{The full license is in \file{\$RHOME/COPYING} and you'd also
  like to look at the R FAQ in 
  \url{http://www.ci.tuwien.ac.at/~hornik/R/R-FAQ.html}.}
\examples{license()}
\keyword{misc}
