\name{logical}
\title{Logical Vectors}
\usage{
logical(n=0)
as.logical(x)
is.logical(x)
}
\alias{logical}
\alias{as.logical}
\alias{is.logical}
\description{
\code{logical} creates a logical vector of the specified length.
Each element of the vector is equal to \code{FALSE}.

\code{as.logical} attempts to coerce its argument to be of
logical type.

\code{is.logical} returns \code{TRUE} or \code{FALSE}
depending on whether its argument is of logical type or not.
}
\keyword{classes}
\keyword{logic}
