\name{median}
\title{Median Value}
\usage{
median(x, na.rm=FALSE)
}
\alias{median}
\description{
\code{median}
computes the sample median of the vector of values given as its argument.
If \code{na.rm} is \code{TRUE} then \code{NA} values are deleted
before computation proceeds.
}
\keyword{univar}
\keyword{robust}
