\name{name}
\title{Variable Names}
\usage{
name(string)
as.name(x)
is.name(x)
}
\alias{name}
\alias{as.name}
\alias{is.name}
\description{
  \code{name} takes the given character string and returns a name (or
  symbol) which has that string as a print-name.

  \code{as.name} coerces its argument to be a name.  The argument must
  be of mode \code{"character"}.

  \code{is.name} returns \code{TRUE} or \code{FALSE} depending on
  whether its argument is a name or not.
}
\keyword{programming}
\keyword{attribute}
