\name{page}
\title{Invoke a Pager on an R Object}
\usage{
page(x)
}
\alias{page}
\arguments{
\item{x}{the name of an \R object.}
}
\description{
    Displays the object named by \code{x} in a pager
}
\author{B. D. Ripley}
\seealso{
    \code{\link{file.show}}, \code{\link{edit}}, \code{\link{fix}}.
}
\keyword{utilities}
