\name{precipitation}
\title{Annual Precipitation in US Cities}
\usage{data(precipitation)}
\alias{precipitation}
\format{A named vector of length 70.

}
\source{Statistical Abstracts of the United States, 1975.

}
\description{The average amount of precipitation (rainfall) in inches for
each of 70 United States (and Puerto Rico) cities.

}
\references{McNeil, D. (1977).  Interactive Data Analysis.  New York: Wiley.

}
\keyword{datasets}
