\name{rivers}
\title{Lengths of Major North American Rivers}
\usage{data(rivers)}
\alias{rivers}
\format{A vector containing 141 observations.

}
\source{World Almanac and Book of Facts, 1975, page 406.

}
\description{This data set gives the lengths  (in miles) of 141 "major" rivers
in North America, as compiled by the US Geological Survey.

}
\references{McNeil, D. (1977).  Interactive Data Analysis.  New York: Wiley.

}
\keyword{datasets}
