\name{rownames}
\alias{rownames}
\alias{rownames<-}
\title{Row Names}
\usage{
rownames(x)
rownames(x) <- namevector
}
\description{
  Retrieve or set the row names of an object (the first component of its
  dimnames).
}
\seealso{
  \code{\link{dimnames}}, \code{\link{colnames}}
}
\keyword{array}
\keyword{manip}
