\name{standardGeneric}
\alias{standardGeneric}
\alias{objWithClass}
\alias{dataClass}
\title{ Formal Method System Placeholders }
\description{
  These routines are primitives used with the \code{methods} package.
  See the documentation for that package.  They cannot be used without
  it and should not be called directly in any case.
}
\usage{
standardGeneric(f)
objWithClass(object, value)
dataClass(object)
}
\author{ John Chambers }
\keyword{programming}
\keyword{internal}
