\name{state}
\alias{state}
\alias{state.abb}
\alias{state.area}
\alias{state.division}
\alias{state.name}
\alias{state.region}
\title{States of the U.S.A.}
\usage{data(state)}
\description{Data sets related to the 50 states of the United States of
  America.}
\details{\R currently contains the following `state' data sets.  Note
  that all data are arranged according to alphabetical order of the
  state names.
  \itemize{
    \item \code{state.abb}: character vector of 2-letter abbreviations
    for the state names.
    \item \code{state.area}: numeric vector of state areas.
    \item \code{state.division}: factor giving state divisions (New
    England, Middle Atlantic, South Atlantic, East South Central, West
    South Central, East North Central, West North Central, Mountain, and
    Pacific).
    \item \code{state.name}: character vector giving the full state
    names
    \item \code{state.region}: factor giving the region (Northeast,
    South, North Central, West) that each state belongs to.
  }
}
\keyword{datasets}
