\name{sum}
\title{Sum of Vector Elements}
\usage{
sum(\dots, na.rm=FALSE)
}
\alias{sum}
\description{
\code{sum} returns the sum of all the values
present in its arguments.
If \code{na.rm} is \code{FALSE} an \code{NA}
value in any of the arguments will cause
a value of \code{NA} to be returned, otherwise
\code{NA} values are ignored.
}
\arguments{
  \item{\dots}{numeric vectors.}
  \item{na.rm}{logical. Should missing values be removed?}
}
\keyword{arith}
