\name{sunspots}
\alias{sunspots}
\title{Monthly Mean Relative Sunspot Numbers}
\usage{data(sunspots)}
\description{Zurich monthly sunspot numbers from 1749 to 1983.}
\format{A time series of 229 cases (years from 1749 to 1983) on 12
    variables (the mean relative sunspot numbers in the months January
    to December).  The last case is incomplete.
}
\source{
    D. F. Andrews and A. M. Herzberg (1985).
    \emph{Data: A Collection of Problems from Many Fields for the
	Student and Research Worker}.
    New York: Springer-Verlag.
}
\keyword{datasets}
