% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-theme.R
\name{theme_omics}
\alias{theme_omics}
\title{The plot theme to be used for figures in the OmicsPlayground app.}
\usage{
theme_omics(
  style = "default",
  base_size = 15,
  grid = "xy",
  axistitle = "xy",
  axistext = "xy",
  axis_num = "none",
  legend_num = FALSE,
  panelborder = FALSE,
  margin = 0,
  ...
)
}
\arguments{
\item{style}{(string) Overall color style of text labels.
Either "default" or "light".}

\item{base_size}{(integer) Base point size}

\item{grid}{(string) Grid lines. Options include  "none" or
any combination of "X", "Y", "x" and "y".}

\item{axistitle}{(string) Axis titles. Options include "none" or
any combination of "X", "Y", "x" and "y".}

\item{axistext}{(string) Axis text labels for values or groups.
Options include "none" or any combination of "X", "Y", "x" and "y".}

\item{axis_num}{(string) Should axis text be formatted as monospaced? Set 
to  x and y, respectively, in case numeric values are displayed. Options 
include "none" or any combination of "X", "Y", "x" and "y".}

\item{legend_num}{(logical) Should legend text be formatted as monospaced?
Defaults to FALSE (no monospace font). Set to TRUE in case of numeric values.
#' @param panelborder (logical) Should a panel border be drawn?
Defaults to FALSE (no border). If TRUE it also adds tick marks to both axes.}

\item{margin}{(numeric) Should a margin of x be added to the plot?
Defaults to 0 (no margin by default).}

\item{...}{Other arguments passed to ggplot methods.}
}
\description{
The plot theme to be used for figures in the OmicsPlayground app.
}
\examples{
\dontrun{
library(ggplot2)
ggplot(mpg, aes(class)) + geom_bar() + theme_omics()
ggplot(mpg, aes(class)) + geom_bar() +
  theme_omics(style = "light", grid = "xy", margin = 20)
ggplot(mpg, aes(class)) + geom_bar() +
  theme_omics(grid = "none", axistext = "x", 
              axistitle = "none", panelborder = TRUE)
}

}
