\name{trees}
\title{Girth, Height and Volume for Black Cherry Trees}
\usage{data(trees)}
\alias{trees}
\format{A data frame with 31 observations on 3 variables.
  \tabular{rlll}{
  [,1]  \tab Girth   \tab numeric  \tab Tree diameter at 4 ft 6 in\cr
  [,2]  \tab Height  \tab numeric  \tab Height in ft\cr
  [,3]  \tab Volume  \tab numeric  \tab Volume of timber in cubic ft\cr
}
\source{T. A. Ryan, B. L. Joiner and B. F. Ryan (1976).
The Minitab Student Handbook. Duxbury Press.

}
\description{This data set provides measurements of the girth, height and volume
of timber in 31 felled black cherry trees.  Note that girth is the
diameter of the tree at 4 ft 6 in above the ground.

}
\references{A. C. Atkinson (1985).  Plots, Transformations and Regression.
Oxford University Press.

}
\keyword{datasets}
