\name{typeof}
\title{The Type of an Object}
\usage{
typeof(x)
}
\alias{typeof}
\description{
  \code{typeof} determines the (\R internal)
  type or storage mode of any object \code{x}.  It returns a character string.
}
\seealso{
  \code{\link{mode}}, \code{\link{storage.mode}}.
}
\examples{
typeof(2)
mode(2)
}
\keyword{attribute}
