% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgx-functions.R
\name{util.findboxes}
\alias{util.findboxes}
\title{Given a Set of Points and Box sizes,
https://github.com/slowkow/ggrepel/issues/24}
\usage{
util.findboxes(
  df,
  xcol,
  ycol,
  box_padding_x,
  box_padding_y,
  point_padding_x,
  point_padding_y,
  xlim,
  ylim,
  force = 1e-07,
  maxiter = 20000
)
}
\description{
Given a Set of Points and Box sizes,
https://github.com/slowkow/ggrepel/issues/24
}
