\name{women}
\title{Average Heights and Weights for American Women}
\usage{data(women)}
\alias{women}
\format{A data frame with 15 observations on 2 variables.
  \tabular{rlll}{
    [,1]  \tab height  \tab numeric  \tab Height (in)\cr
    [,2]  \tab weight  \tab numeric  \tab Weight (lbs)
    }
}
\source{The World Almanac and Book of Facts, 1975.

}
\description{This data set gives the average heights and weights for American
women aged 30-39.

}
\references{McNeil, D. R. (1977).  Interactive Data Analysis.  Wiley.

}
\keyword{datasets}
