% File src/library/base/man/zutils.Rd
% Part of the R package, http://www.R-project.org
% Copyright 1995-2007 R Core Team
% Distributed under GPL 2 or later

\name{zutils}
\alias{.standard_regexps}
\title{Miscellaneous Internal/Programming Utilities}
\description{Miscellaneous internal/programming utilities.}
\usage{
.standard_regexps()
}
\details{
  \code{.standard_regexps} returns a list of \sQuote{standard} regexps,
  including elements named \code{valid_package_name} and
  \code{valid_package_version} with the obvious meanings.  The regexps
  are not anchored.
}
\keyword{utilities}
