\name{attitude}
\alias{attitude}
\title{Attitudes Toward Supervisors}
\usage{data(attitude)}
\description{Not available (yet).}
\format{A data from with 30 observations on 7 variables named
    \code{rating} (Overall rating), \code{complaints} (Handling of
    employee complaints), \code{privileges} (Gives special privileges),
    \code{learning} (Opportunity to learn), \code{raises} (Gives
    raises), \code{critical} (Too critical), and \code{advance}
    (Advancement).
}
\source{
    S. Chatterjee and B. Price (1977).
    \emph{Regression Analysis by Example}.
    New York: Wiley.
}
\keyword{datasets}
