\name{colnames}
\alias{colnames}
\alias{colnames<-}
\title{Column Names}
\usage{
colnames(x)
colnames(x) <- namevector
}
\description{
  Retrieve or set the column names of an object (the second component of
  its dimnames).
}
\seealso{
  \code{\link{dimnames}}, \code{\link{rownames}}
}
\keyword{array}
\keyword{manip}
