\name{contributors}
\title{R Project Contributors}
\alias{contributors}
\section{The \R Project}{
  \R is a project which is attempting to provide a modern
  piece of statistical software to the GNU suite of software.
}
\section{Contributors}{
  The current \R is the result of a collaborative effort
  authors from all over the world.
  \R was initially written by Robert Gentleman and Ross Ihaka ---
  also known as ``\emph{R & R}'' ---
  of the Statistics Department of the University of Auckland.

  Since mid-1997 there has been a core group with access to the \R
  source, currently consisting of

  \itemize{%%-- should we add E-mail  and / or affiliation ?
      \item Douglas Bates
      \item Peter Dalgaard
      \item Robert Gentleman
      \item Kurt Hornik
      \item Ross Ihaka
      \item Friedrich Leisch
      \item Thomas Lumley
      \item Martin Maechler
      \item Paul Murrell
      \item Heiner Schwarte
      \item Luke Tierney
    }

  In addition, a large group of individuals has contributed to \R by
  donating code, bug reports and documentation, notably

  Valerio Aimalea,
  Ben Bolker,
  Goran Brostrom,
  Paul Gilbert,
  Robert King,
  Philippe Lambert,
  Patrick Lindsey,
  Jim Lindsey,
  Guido Masarotto,
  Steve Oncley and Gordon Maclean,
  Richard Okeefe,
  Hubert Palme,
  Martyn Plummer,
  Brian Ripley,
  Bill Simpson,
  Bill Venables, and
  Andreas Weingessel.

  We have probably omitted some important names here because of
  incomplete record keeping.  If we have overlooked you, please
  let us know and we'll update the list.

  A special debt is owed to John Chambers who has graciously
  contributed advice and encouragement.
}
\keyword{misc}
