\name{contributors}
\alias{contributors}
\title{R Project Contributors}
\description{
  The \R Who-is-who.
}
\section{The R Project}{
  \R is a project which is attempting to provide a modern piece of
  statistical software for the GNU suite of software.
}
\section{Contributors}{
  The current \R is the result of a collaborative effort with
  contributions from all over the world.

  \R was initially written by Robert Gentleman and Ross Ihaka---also
  known as ``\emph{R & R}''---of the Statistics Department of the
  University of Auckland.

  Since mid-1997 there has been a core group with write access to the \R
  source, currently consisting of

  \itemize{%%-- should we add E-mail  and / or affiliation ?
    \item Douglas Bates
    \item John Chambers
    \item Peter Dalgaard
    \item Robert Gentleman
    \item Kurt Hornik
    \item Ross Ihaka
    \item Friedrich Leisch
    \item Thomas Lumley
    \item Martin Maechler
    \item Guido Masarotto
    \item Paul Murrell
    \item Brian Ripley
    \item Duncan Temple Lang
    \item Luke Tierney
  }

  plus Heiner Schwarte up to October 1999.

  In addition, a large group of individuals has contributed to \R by
  donating code, bug reports and documentation, notably

  Valerio Aimale,
  Thomas Baier,
  Ben Bolker,
  Goran Brostrom,
  Saikat DebRoy,
  Lyndon Drake,
  Paul Gilbert,
  Robert King,
  Philippe Lambert,
  Patrick Lindsey,
  Jim Lindsey,
  John Maindonald,
  Jens Oehlschaegel-Akiyoshi,
  Steve Oncley and Gordon Maclean,
  Richard Okeefe,
  Hubert Palme,
  Jose Pinheiro,
  Martyn Plummer,
  Jonathan Rougier,
  Bill Simpson,
  Adrian Trapletti,
  Terry Therneau,
  Bill Venables,
  Gregory R. Warnes,
  and
  Andreas Weingessel.

  We have probably omitted some important names here because of
  incomplete record keeping.  If we have overlooked you, please let us
  know and we'll update the list.

  A special debt is owed to John Chambers who has graciously contributed
  advice and encouragement in the early days of \R and later became
  a member of the core team.
}
\keyword{misc}
