\name{Defunct}
\title{Defunct Functions}
\alias{Defunct}
\alias{.Defunct}
\usage{
.Defunct()

%--- oldfun(.)
}
%------ PLEASE: put \alias{.} here for  EACH !
%\alias{...}
%
\description{
  The functions listed here are no longer part of \R as they are not
  needed any more.

  \code{.Defunct} is the function to which defunct functions are set.
}
\seealso{\code{\link{Deprecated}}}
\keyword{documentation}
\keyword{utilities}

