\name{Deprecated}
%------ NOTE:  ../R/Deprecated.R   must be synchronized with this!
\title{Deprecated Functions}
\alias{Deprecated}
\alias{.Deprecated}
%------ PLEASE: put \alias{.} here for EACH !
\description{
  These functions are provided for compatibility with older versions of
  \R only, and may be defunct as soon as of the next release.
}
\usage{
.Deprecated(new)

}
\details{
  \code{.Deprecated("<new name>")} is called from deprecated functions.
}
\seealso{\code{\link{Defunct}}}
\keyword{documentation}
\keyword{utilities}
