\name{Deprecated}
\title{Deprecated Functions}
\alias{Deprecated}
\usage{
dnchisq
pnchisq
qnchisq
print.plot
save.plot
}
\description{
  These functions are provided for compatibility with older versions of
  \R{} only, and may be defunct as soon as of the next release.

  The \code{*chisq()} functions now take an optional non-centrality
  argument, so the \code{*nchisq()} functions are no longer needed.

  The new function \code{dev.print()} should now be used for saving
  plots to a file or printing them.
}
\seealso{\code{\link{Defunct}}}
