\name{eurodist}
\title{Distances Between Cities in Europe}
\usage{data(eurodist)}
\alias{eurodist}
\format{A dist object based on 21 objects.  (You must have the mva library
loaded to have the methods for this kind of object available).
}
\source{Crystal, D. Ed. (1990).  ``The Cambridge Encyclopaedia''.
Cambridge: Cambridge University Press,

}
\description{The data give the road distances (in km) between 21 cities in Europe.
  The data are taken from a table in ``The Cambridge Encyclopaedia''.
}
\keyword{datasets}
