\name{file.path}
\alias{file.path}
\title{Construct Path to File}
\description{
 Construct the path to a file from components in a platform-independent way.
}
\usage{
file.path(..., fsep=.Platform$file.sep)
}
\arguments{
 \item{\dots}{Character vectors}
 \item{fsep}{The path separator to use}
}
\value{
  A character vector of length one.
}
\keyword{file}
