\name{frame}
\title{Start a New Plot Frame}
\usage{
frame()
}
\alias{frame}
\value{
The function is invoked for its side effect which is to
cause the completion of plotting in the current plot
and an advance to a new graphics frame.
This is useful for skipping plots when a multi-figure region is in use.
}
\keyword{iplot}
