\name{graphics.off}
\title{Turn Off All Graphics Devices}
\usage{
dev.off()
graphics.off()
}
\alias{dev.off}
\alias{graphics.off}
\description{
These two functions are symonyms and shutdown the current graphics device.

In the future, with \emph{multiple} concurrent devices, things will change
\ldots.
}
\seealso{
\code{\link{Devices}}.
}
\keyword{device}
\keyword{iplot}
