\name{is.single}
\title{Is an Object of Single Precision Type?}
\usage{is.single(x)}
\alias{is.single}
\description{
\code{is.single} reports an error. There are no single precision
values in R.
}
\arguments{
  \item{x}{object to be tested.}
}
\details{
  \code{is.single} is generic: you can write methods to handle
  of specific classes of objects, see \link{InternalMethods}.
}
\keyword{classes}
