\name{license}
\title{Show the License Terms for R}
\usage{license()}
\alias{license}
\description{
The function just prints a few lines telling about the GNU license
terms, and how to obtain more details.
}
\references{The full license is in \code{$RHOME/COPYING}.}
\examples{license()}
\keyword{misc}
