\name{license}
\title{The R License Terms}
\alias{license}
\alias{licence}
\description{
  The license terms under which \R is distributed.
}
\usage{
license()
licence()
}
\details{
  \R is distributed under the terms of the GNU GENERAL PUBLIC LICENSE
  Version 2, June 1991.  A copy of this license is in file
  \file{\$R\_HOME/COPYING} and can be viewed by \code{RShowDoc("COPYING")}.

  A small number of files (the API header files)   are distributed under
  the LESSER GNU GENERAL PUBLIC LICENSE version 2.1. A copy of this
  license is in file \file{\$R\_DOC\_DIR/COPYING.LIB} and can be viewed
  by \code{RShowDoc("COPYING.LIB")}.
}
\keyword{misc}
