% File src/library/base/man/license.Rd
% Part of the R package, https://www.R-project.org
% Copyright 1995-2013 R Core Team
% Distributed under GPL 2 or later

\name{license}
\title{The R License Terms}
\alias{license}
\alias{licence}
\description{
  The license terms under which \R is distributed.
}
\usage{
license()
licence()
}
\details{
  \R is distributed under the terms of the GNU GENERAL PUBLIC LICENSE,
  either Version 2, June 1991 or Version 3, June 2007.  A copy of the
  version 2 license is in file \file{\var{\link{R_HOME}}/doc/COPYING}
  and can be viewed by \code{RShowDoc("COPYING")}.  Version 3 of the
  license can be displayed by \code{RShowDoc("GPL-3")}.

  A small number of files (some of the API header files) are distributed
  under the LESSER GNU GENERAL PUBLIC LICENSE, version 2.1 or later.  A
  copy of this license is in file \file{$R_SHARE_DIR/licenses/LGPL-2.1}
  and can be viewed by \code{RShowDoc("LGPL-2.1")}.  Version 3 of the
  license can be displayed by \code{RShowDoc("LGPL-3")}.
}
\keyword{misc}
