\name{nlevels}
\title{The Number of Levels of a Factor}
\usage{
nlevels(x)
}
\alias{nlevels}
\description{
\code{nlevels}
returns the number of levels which its argument has.
If the argument is not a factor \code{NA} is returned.

The actual factor levels (if they exist) can be obtained
with the \code{levels} function.
}
\seealso{
\code{\link{levels}}.
}
\keyword{category}
