\name{offset}
\title{Include an offset in a model formula}
\usage{
offset(x)
}
\alias{offset}
\arguments{
 \item{x}{An offset to be included in a model frame}
}
\description{
An offset is a term to be added to a linear predictor, such as in a generalised linear model, with known coefficient 1 rather than an estimated coefficient.
}
\value{
The input value.
}


\seealso{\code{\link{model.offset}}, \code{\link{model.frame}}, \code{\link{glm}}}

\keyword{models}
