\name{presidents}
\title{Approval Rating of US Presidents}
\usage{data(presidents)}
\alias{presidents}
\format{A time series of 120 values.

}
\source{The Gallup Organisation.

}
\description{The (approximately) quarterly approval rating for the President
of the United states from the first quarter of 1945 to the last
quarter of 1974.

The data are actually a fudged version of this.  See McNeil's
book for details.

}
\references{McNeil, D. (1977).  Interactive Data Analysis.  New York: Wiley.

}
\keyword{datasets}
