\name{quantile}
\title{Sample Quantiles}
\usage{
quantile(x, probs=seq(0, 1, 0.25), na.rm=FALSE)
}
\alias{quantile}
\description{
\code{quantile}
produces sample quantiles corresponding to the
given probabilities.
The smallest observation corresponds to
a probability of 0 and the largest to a probability of 1.
}
\keyword{univar}
