\name{Reserved}
\alias{Reserved}
\alias{reserved}
\title{Reserved Words in R}
\description{
  The reserved words in \R are

  \code{\link{if}} \code{\link{else}} \code{\link{repeat}} \code{\link{while}}
  \code{\link{function}} \code{\link{for}} \code{in}
  \code{\link{next}} \code{\link{break}}

  \code{\link{TRUE}} \code{\link{FALSE}} \code{\link{NULL}}
  \code{\link{Inf}} \code{\link{NaN}}
  \code{\link{NA}} \code{\link{NA_integer_}} \code{\link{NA_real_}}
  \code{\link{NA_complex_}} \code{\link{NA_character_}}

  \code{\dots} and \code{..1}, \code{..2} etc, used to refer to
  arguments passed down from an enclosing function.
}
\keyword{programming}
\keyword{documentation}
