\name{ToothGrowth}
\title{The Effect of Vitamin C on Tooth Growth in Guinea Pigs}
\usage{data(ToothGrowth)}
\alias{ToothGrowth}
\format{A data frame with 60 observations on 3 variables.
    \tabular{rlll}{
    	[,1]  \tab len   \tab numeric  \tab Tooth length\cr
    	[,2]  \tab supp  \tab factor   \tab Supplement (VC or OJ).\cr
    	[,3]  \tab dose  \tab numeric  \tab Dose in milligrams.
    }
}
\source{
    C. I. Bliss (1952).  The Statistics of Bioassay, Academic Press.
}
\description{
    The response is the average length of odontoblasts (teeth) in each
    of 10 guinea pigs.  Three dose levels 0.5, 1, and 2 mg of juice and
    acid in the juice were used.

    This makes no sense - check the original source.
}
\references{
    McNeil, D. R. (1977).  Interactive Data Analysis.  New York: Wiley.
}
\keyword{datasets}
