\name{typeof}
\title{The Type of an Object}
\usage{
typeof(x)
}
\alias{typeof}
\description{
\code{typeof} provides a means of determining the
type or storage mode of an object.
\code{typeof} returns a character string giving the type
of the object.
}
\seealso{
\code{\link{mode}}, \code{\link{storage.mode}}.
}
\examples{
typeof(2)
mode(2)
}
\keyword{attribute}
