% File src/library/base/man/units.Rd
% Part of the R package, https://www.R-project.org
% Copyright 2023 R Core Team
% Distributed under GPL 2 or later

\name{units}
\title{Units}
\alias{units}
\alias{units<-}
\description{Get or set units.}
\usage{
units(x)
units(x) <- value
}
\arguments{
  \item{x}{an \R object}
  \item{value}{an \R object}
}
\details{
  These are generic functions, with methods for \code{"\link{difftime}"}
  objects.
}
