\name{zutils}
\alias{.standard_regexps}
\title{Miscellaneous Internal/Programming Utilities}
\description{Miscellaneous internal/programming utilities.}
\usage{
.standard_regexps()
}
\details{
  \code{.standard_regexps} returns a list of \dQuote{standard} regexps,
  including elements named \code{valid_package_name} and
  \code{valid_package_version} with the obvious meanings.  The regexps
  are not anchored.
}
\keyword{utilities}
